refactor(code-mode): rename "harness" to "runner"
The sandbox bootstrap module was called "harness", which is overloaded
(test-harness connotations) and jargon-y. Rename to "runner", which plainly
describes the script that runs the user's program inside the sandbox and
bridges RPC over stdio.

- harness.ts -> runner.ts
- buildHarness -> buildRunner, BuildHarnessOptions -> BuildRunnerOptions
- CodeModeHarnessFrame -> CodeModeRunnerFrame
- updated comments and exports

No behavior change; 17 code-mode tests pass, typecheck and lint clean.
