+++ packages/core/src/tools/code-mode/index.ts
@@ -1,13 +1,13 @@
 export { createCodeMode, createCodeModeTool, type CodeModeResult } from './code-mode';
 export { createCodeModeInstructions, generateStubs, jsonSchemaToTsString, type CodeModeStub } from './stub-generator';
 export { StdioCodeModeTransport } from './transport';
-export { buildHarness, buildProgramModule, FRAME_PREFIX } from './harness';
+export { buildRunner, buildProgramModule, FRAME_PREFIX } from './runner';
 export type {
   CodeModeConfig,
   CodeModeToolResult,
   CodeModeTransport,
   CodeModeToolDispatcher,
-  CodeModeHarnessFrame,
+  CodeModeRunnerFrame,
   CodeModeRpcRequest,
   CodeModeRpcResponse,
   CodeModeLogEvent,
